combine writes into bigger batches
Previously we flush the zip archive on each entry. This is slow for many files.
Now, we add up to 10MiB of buffers to the archive, moving ownership to
libzip, before flushing and causing libzip to free those buffers.

This is based on functionality recently added to axfive_libzip:
https://gitlab.com/Taywee/axfive-libzip/-/merge_requests/4

Hint: The diff is best viewed without whitespace changes.